requirements.txt
customtkinter>=5.2.0
bcrypt>=4.0.1
python-docx>=0.8.11
tkinter 